Adding net new functions and modifying/deprecating existing ones
Created a catch-all function to pull down relevant data points for a Tale of the Tape style breakdown for a given team. Next steps is to test data upload to PostgreSQL db, and coming up with a way to pull the next opponent for a given team and pull their TOTT (Tale of the Tape) data points.